requirements.txt
Babel==2.6.0
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-ckeditor==5.6.1
django-extra-views==0.11.0
django-haystack==2.8.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-oscar==1.6.7
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.3
factory-boy==2.11.1
Faker==1.0.5
gunicorn==19.9.0
mock==2.0.0
pbr==5.1.3
phonenumbers==8.10.9
Pillow==6.0.0
psycopg2==2.7.7
purl==1.5
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
sorl-thumbnail==12.4.1
text-unidecode==1.2
Unidecode==1.0.23
whitenoise==4.1.2
